requirements.txt
fastapi==0.47.1
uvicorn==0.11.2
pyjwt==1.7.1
python-dotenv==0.10.5
